requirements.txt
pandas>=1.4.0
pytest>=6.2.0
pyyaml>=6.0